requirements.txt
timm
tqdm
numpy
einops
pyyaml
json
matplotlib
python-hostlist
sk-video
torch==1.11.0
torchvision==0.12.0
wandb==0.12.16
